Compose: inject source volumes
This is huge for developers. It means that the source code resides in the
developer machine but the same contents are shared with the container making
possible to see the results and get feedback from the environment instantly.

If you're running Docker natively this change shouldn't cause any performance
issue. On the other hand, if you're using Docker for Mac or similar please be
aware that the host filesystem is shared with the virtual machine running
Linux. Fortunately, Docker hides that complexity from the user so extra config
is not required but you may still be affected by certain level of performance
degradation, particularly when sharing large folders.

Gunicorn (used in Dashboard and Storage Service) is deployed with the reload
engine enabled, which uses inotify to efficiently detect code changes and
reload the workers in the pool. This means that you can edit a file from your
favorite IDE, save, go back to your browser, refresh and see the results right
away.