Shibboleth: added am-shib compose configuration, which presents Archivematica with Shibboleth auth
We don't protect api and static resources in nginx and SP config because the Archivematica APIs are both already protected because they require a valid username and API key to access them.

Static resources, /media for Dashboard and /static for Storage Service, don't require protection because they are not sensitive data.